Import project from Eclipse to IntelliJ and Migrating from Java 7 to Java 8 Runtime. fix #459 @2.0h #255 @1.0h
