--- src/META-INF/jdoconfig.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<jdoconfig xmlns="http://java.sun.com/xml/ns/jdo/jdoconfig"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:noNamespaceSchemaLocation="http://java.sun.com/xml/ns/jdo/jdoconfig">
-
-	<persistence-manager-factory name="transactions-optional">
-		<!-- https://cloud.google.com/appengine/docs/java/datastore/jdo/overview-dn2 -->
-		<property name="javax.jdo.PersistenceManagerFactoryClass"
-			value="org.datanucleus.api.jdo.JDOPersistenceManagerFactory" />
-		<property name="javax.jdo.option.ConnectionURL" value="appengine" />
-		<property name="javax.jdo.option.NontransactionalRead" value="true" />
-		<property name="javax.jdo.option.NontransactionalWrite"
-			value="true" />
-		<property name="javax.jdo.option.RetainValues" value="true" />
-		<property name="datanucleus.appengine.autoCreateDatastoreTxns"
-			value="true" />
-	</persistence-manager-factory>
-</jdoconfig>
