--- src/com/appspot/OIT_Maikata_Fan/PMF.java
@@ -1,16 +1,0 @@
-package com.appspot.OIT_Maikata_Fan;
-
-import javax.jdo.JDOHelper;
-import javax.jdo.PersistenceManagerFactory;
-
-public class PMF {
-	private static final PersistenceManagerFactory pmfInstance = JDOHelper
-			.getPersistenceManagerFactory("transactions-optional");
-
-	private PMF() {
-	}
-
-	public static PersistenceManagerFactory get() {
-		return pmfInstance;
-	}
-}
