+++ src/main/webapp/WEB-INF/appengine-web.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  Copyright 2016 Google Inc.
+  Licensed under the Apache License, Version 2.0 (the "License");
+  you may not use this file except in compliance with the License.
+  You may obtain a copy of the License at
+        http://www.apache.org/licenses/LICENSE-2.0
+  Unless required by applicable law or agreed to in writing, software
+  distributed under the License is distributed on an "AS IS" BASIS,
+  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  See the License for the specific language governing permissions and
+  limitations under the License.
+-->
+<!-- [START config] -->
+<!-- https://cloud.google.com/appengine/docs/standard/java/config/appref -->
+<appengine-web-app xmlns="http://appengine.google.com/ns/1.0">
+    <threadsafe>false</threadsafe>
+    <runtime>java8</runtime>
+    <url-stream-handler>urlfetch</url-stream-handler>
+</appengine-web-app>
+        <!-- [END config] -->
