--- war/WEB-INF/appengine-generated/datastore-indexes-auto.xml
@@ -1,4 +1,0 @@
-<!-- Indices written at Sun, 12 May 2013 14:51:40 JST -->
-
-<datastore-indexes/>
-
