--- war/WEB-INF/appengine-web.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<appengine-web-app xmlns="http://appengine.google.com/ns/1.0">
-	<application>oit-maikata-fan</application>
-	<version>1</version>
-
-	<!-- Configure java.util.logging -->
-	<system-properties>
-		<property name="java.util.logging.config.file" value="WEB-INF/logging.properties" />
-
-		<!-- https://sites.google.com/site/elekmole/gaebottop/gaebotbasic/06-set-consumerkey -->
-		<!-- <property name="oauth.consumerKey" value="73FKZ1VWMYze0ke59VnV4g" 
-			/> <property name="oauth.consumerSecret" value="ZYxxqsWJJkYWFpo7zxf431JT0pgctbGPacLa6Zig" 
-			/> -->
-	</system-properties>
-
-	<!-- https://sites.google.com/site/elekmole/gaebottop/gaebotbasic/06-set-consumerkey -->
-	<sessions-enabled>true</sessions-enabled>
-
-	<threadsafe>false</threadsafe>
-
-</appengine-web-app>
