--- war/WEB-INF/web.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns="http://java.sun.com/xml/ns/javaee" xmlns:web="http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd"
-	xsi:schemaLocation="http://java.sun.com/xml/ns/javaee
-http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd"
-	version="2.5">
-	<servlet>
-		<servlet-name>CalledEvery15MinsServlet</servlet-name>
-		<servlet-class>com.appspot.OIT_Maikata_Fan.CalledEvery15MinsServlet</servlet-class>
-	</servlet>
-	<servlet-mapping>
-		<servlet-name>CalledEvery15MinsServlet</servlet-name>
-		<url-pattern>/cron/every15mins</url-pattern>
-	</servlet-mapping>
-
-	<servlet>
-		<servlet-name>CalledEvery1630Servlet</servlet-name>
-		<servlet-class>com.appspot.OIT_Maikata_Fan.CalledEvery1630Servlet</servlet-class>
-	</servlet>
-	<servlet-mapping>
-		<servlet-name>CalledEvery1630Servlet</servlet-name>
-		<url-pattern>/cron/every1630</url-pattern>
-	</servlet-mapping>
-
-	<servlet>
-		<servlet-name>TestServlet</servlet-name>
-		<servlet-class>com.appspot.OIT_Maikata_Fan.TestServlet</servlet-class>
-	</servlet>
-	<servlet-mapping>
-		<servlet-name>TestServlet</servlet-name>
-		<url-pattern>/cron/test</url-pattern>
-	</servlet-mapping>
-
-	<!-- https://developers.google.com/appengine/docs/java/config/cron?hl=ja -->
-	<security-constraint>
-		<web-resource-collection>
-			<web-resource-name>admin page</web-resource-name>
-			<url-pattern>/cron/*</url-pattern>
-		</web-resource-collection>
-		<auth-constraint>
-			<role-name>admin</role-name>
-		</auth-constraint>
-	</security-constraint>
-
-	<servlet>
-		<servlet-name>SystemServiceServlet</servlet-name>
-		<servlet-class>com.google.api.server.spi.SystemServiceServlet</servlet-class>
-		<init-param>
-			<param-name>services</param-name>
-			<param-value />
-		</init-param>
-	</servlet>
-	<servlet-mapping>
-		<servlet-name>SystemServiceServlet</servlet-name>
-		<url-pattern>/_ah/spi/*</url-pattern>
-	</servlet-mapping>
-</web-app>
-
