--- war/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
-<!-- The HTML 4.01 Transitional DOCTYPE declaration-->
-<!-- above set at the top of the file will set     -->
-<!-- the browser's rendering engine into           -->
-<!-- "Quirks Mode". Replacing this declaration     -->
-<!-- with a "Standards Mode" doctype is supported, -->
-<!-- but may lead to some differences in layout.   -->
-
-<html>
-<head>
-<meta http-equiv="content-type" content="text/html; charset=UTF-8">
-<title>大阪工業大学枚方ファン</title>
-</head>
-
-<body>
-  <h1>大阪工業大学枚方ファン</h1>
-  大阪工業大学枚方キャンパスの非公式アカウント．まいかた勢向けの情報を自動ツイート．
-  <h2>仕様</h2>
-  毎日以下の時間に自動ツイート．
-  <br />・9時
-  <br />・12時30分
-  <br />・16時30分
-  <h2>アカウント</h2>
-  <a href="https://twitter.com/OIT_Maikata_Fan">@OIT_Maikata_Fan</a>
-
-</body>
-</html>
